pyproject.toml
[project]
name = "shedulegeneratorweb"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite~=0.21.0",
    "bcrypt~=4.0.1",
    "fastapi>=0.109.1",
    "jinja2>=3.1.0",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]~=1.7.4",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]~=3.4.0",
    "python-multipart>=0.0.6",
    "starlette>=0.49.1",
    "uvicorn[standard]>=0.24.0",
]
